+++ mex/backend/auxiliary/wikidata.py
@@ -68,7 +68,7 @@ def _fetch_or_insert_organization(name: str) -> ExtractedOrganization:
     if not extracted_item:
         raise NoResultFoundError(name)
     connector = GraphConnector.get()
-    deque(connector.ingest_extracted_items([extracted_item]))
+    deque(connector.ingest_items([extracted_item]))
     logger.info("ingested organization %s", name)
     return extracted_item