+++ CHANGELOG.md
@@ -9,15 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- ingest endpoint now supports rule-sets
 - new endpoint returns merged person from ldap login information
 
 ### Changes
 
 - ldap search endpoint now also returns contact points
+- change rule set ingestion to use v2-style queries
 
 ### Deprecated
 
 ### Removed
+
+- removed black dependency in favor of unittest.mock
 
 ### Fixed